backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
channels==4.0.0
channels-redis==4.1.0
python-binance==1.0.19
websockets==12.0
celery==5.3.4
redis==5.0.1
python-decouple==3.8
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
ta==0.11.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
